+++ src/day_18/xy.rs
@@ -1,11 +1,11 @@
 #[derive(Debug, PartialEq, Eq, Clone, Copy)]
 pub struct Xy {
-    pub x: i32,
-    pub y: i32,
+    pub x: i64,
+    pub y: i64,
 }
 
 impl Xy {
-    pub fn new(x: i32, y: i32) -> Self {
+    pub fn new(x: i64, y: i64) -> Self {
         return Xy { x, y };
     }
 
@@ -13,7 +13,7 @@ impl Xy {
         return Xy::new(self.x + other.x, self.y + other.y);
     }
 
-    pub fn abs_diff(&self, other: &Xy) -> u32 {
+    pub fn abs_diff(&self, other: &Xy) -> u64 {
         return self.x.abs_diff(other.x) + self.y.abs_diff(other.y);
     }
 }
